api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.0

# CPU-only PyTorch (avoid 700MB+ CUDA downloads)
# Note: Python 3.13 requires torch 2.6.0+ and torchvision 0.21.0+
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.5.1
torchvision>=0.21.0

# Sentence transformers (will use CPU torch above)
sentence-transformers>=3.0.0
transformers>=4.30.0

# Vector database
sqlite-vec==0.1.1
aiosqlite>=0.20.0

# Document processing
python-docx>=1.1.2
markdown==3.5.2
python-multipart==0.0.9
watchdog==4.0.0
numpy>=1.26.4,<2.0.0

# Advanced document parsing (default)
# Note: Docling has heavy dependencies, install takes 3-5 min
# Note: HybridChunker available in docling-core 2.8.0+ (requires docling 2.9.0+)
docling>=2.9.0

# Code parsing with AST-based chunking
# Note: Uses tree-sitter for language-agnostic AST parsing
astchunk>=0.1.0

# Tree-sitter language pack for R and other languages (160+ languages supported)
# Note: Actively maintained replacement for tree-sitter-languages
tree-sitter-language-pack>=0.10.0

# Jupyter notebook processing
# Note: nbformat is the official Jupyter format library
nbformat>=5.9.0

# Obsidian vault processing with Graph-RAG
# Note: obsidiantools for wikilinks, tags, frontmatter parsing
# Note: networkx for knowledge graph construction and traversal
obsidiantools>=0.10.0
networkx>=3.0

# Pipeline configuration
PyYAML>=6.0  # YAML config file parsing

# Security - Malware detection
clamd>=1.0.2  # ClamAV daemon client
yara-python>=4.5.0  # YARA pattern matching

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
httpx>=0.27.0  # For TestClient (FastAPI testing)
